Move parsing out of apply/regenerate modules
Now have third-level crate::apply_patches::email and
crate::regenerate_patches::patch_file modules

Avoid usage of mod.rs (no longer needed)
